+++ lib/action_controller/responder.rb
@@ -200,7 +200,7 @@ module ActionController #:nodoc:
       if get?
         raise error
       elsif has_errors? && default_action
-        render :action => default_action
+        render rendering_options
       else
         redirect_to navigation_location
       end
@@ -296,6 +296,14 @@ module ActionController #:nodoc:
 
     def response_overridden?
       @default_response.present?
+    end
+
+    def rendering_options
+      if options[:render]
+        options[:render]
+      else
+        { action: default_action }
+      end
     end
   end
 end